+++ include/igl/eytzinger_aabb.h
@@ -6,6 +6,7 @@
 // v. 2.0. If a copy of the MPL was not distributed with this file, You can 
 // obtain one at http://mozilla.org/MPL/2.0/.
 #ifndef IGL_EYTZINGER_AABB_H
+#define IGL_EYTZINGER_AABB_H
 #include "igl_inline.h"
 #include <Eigen/Core>
