+++ include/igl/eytzinger_aabb_sdf.h
@@ -1,4 +1,12 @@
+// This file is part of libigl, a simple c++ geometry processing library.
+// 
+// Copyright (C) 2025 Alec Jacobson <alecjacobson@gmail.com>
+// 
+// This Source Code Form is subject to the terms of the Mozilla Public License 
+// v. 2.0. If a copy of the MPL was not distributed with this file, You can 
+// obtain one at http://mozilla.org/MPL/2.0/.
 #ifndef IGL_EYTZINGER_AABB_SDF_H
+#define IGL_EYTZINGER_AABB_SDF_H
 #include "igl_inline.h"
 #include <Eigen/Core>
 
@@ -56,4 +64,3 @@ namespace igl
 #  include "eytzinger_aabb_sdf.cpp"
 #endif
 #endif 
-
