+++ include/igl/matlab_format.cpp
@@ -8,6 +8,8 @@
 #include "matlab_format.h"
 #include "STR.h"
 #include "find.h"
+#include "list_to_matrix.h"
+
 
 template <typename DerivedM>
 IGL_INLINE const Eigen::WithFormat< DerivedM > igl::matlab_format(
@@ -38,6 +40,17 @@ IGL_INLINE std::string igl::matlab_format_index(
   const Eigen::MatrixBase<DerivedM> & M,
   const std::string name)
 {
+  // can't return WithFormat since that uses a pointer to matrix
+  return STR(igl::matlab_format((M.array()+1).eval(),name));
+}
+
+template <typename T>
+IGL_INLINE std::string igl::matlab_format_index(
+  const std::vector<T> & Mvec,
+  const std::string name)
+{
+  Eigen::Matrix<T,Eigen::Dynamic,1> M;
+  igl::list_to_matrix(Mvec,M);
   // can't return WithFormat since that uses a pointer to matrix
   return STR(igl::matlab_format((M.array()+1).eval(),name));
 }
@@ -118,6 +131,7 @@ IGL_INLINE const std::string igl::matlab_format(
 
 #ifdef IGL_STATIC_LIBRARY
 // Explicit template instantiation
+template std::basic_string<char, std::char_traits<char>, std::allocator<char>> igl::matlab_format_index<int>(std::vector<int, std::allocator<int>> const&, std::basic_string<char, std::char_traits<char>, std::allocator<char>>);
 // generated by autoexplicit.sh
 template std::basic_string<char, std::char_traits<char>, std::allocator<char> > const igl::matlab_format<bool>(Eigen::SparseMatrix<bool, 0, int> const&, std::basic_string<char, std::char_traits<char>, std::allocator<char> >);
 // generated by autoexplicit.sh
